+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## next release
 - Improved automated testing
 - Added an information notice about the processing time of deleting an administrator
+- Ability to change the expiration date of a coupon after its creation
 - Fix a bug: unable to run rake db:migrate on first install
 - Fix a bug: unable to create or edit a coupon of type 'percentage'